Update to error handling
If report is not found we now send back 404. If the token is invalid
(i.e. authorization is invalid) we send back a 401. Otherwise, for all
other errors it is 500. In addition now, when we create the new token
in /tokens/refresh the response code is 201. This brings all
error/success HTTP status code handling in line with the AVAST spec.

In addition we now use 'reason' instead of 'message' within the
resultant error related JSON.